réglage accès pages + changment nom pp de username à id

+++ html/functions/removeFromNewsletter.php
@@ -2,11 +2,14 @@
 
     require 'functions.php';
 
+    if(!isConnected()){
+        die(header("Location: ../index.php"));
+    }
+
     $emailid = getUserId();
 
 	$pdo = connectDB();
     $queryPrepared = $pdo->prepare("DELETE FROM utrackpa_newsletter WHERE emailid='$emailid';");
     $queryPrepared->execute();
 
-    header("Location: ../templates/Home/dash-board.php")
-?>
+        header("Location: ../templates/Home/dash-board.php");
